+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/pojo/template/v2/PipelineTemplateMarketCreateReq.kt
@@ -14,7 +14,7 @@ data class PipelineTemplateMarketCreateReq(
     @get:Schema(title = "研发商店模板项目ID", required = true)
     val marketTemplateProjectId: String,
     @get:Schema(title = "研发商店模板版本", required = true)
-    val marketTemplateVersion: Long
+    val marketTemplateVersion: Int
 ) : PipelineTemplateBasicCreateReq(
     projectId = projectId,
     source = source
