+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/template/v2/UserPipelineTemplateInstanceV2ResourceImpl.kt
@@ -45,7 +45,7 @@ class UserPipelineTemplateInstanceV2ResourceImpl(
         userId: String,
         projectId: String,
         templateId: String,
-        version: Long,
+        version: Int,
         useTemplateSettings: Boolean,
         request: PipelineTemplateInstancesReleaseRequest
     ): TemplateOperationRet {
@@ -63,7 +63,7 @@ class UserPipelineTemplateInstanceV2ResourceImpl(
         userId: String,
         projectId: String,
         templateId: String,
-        version: Long,
+        version: Int,
         useTemplateSettings: Boolean,
         request: PipelineTemplateInstancesReleaseRequest
     ): Result<String> {
@@ -83,7 +83,7 @@ class UserPipelineTemplateInstanceV2ResourceImpl(
         userId: String,
         projectId: String,
         templateId: String,
-        version: Long,
+        version: Int,
         useTemplateSettings: Boolean,
         instances: List<TemplateInstanceUpdate>
     ): TemplateOperationRet {
@@ -101,7 +101,7 @@ class UserPipelineTemplateInstanceV2ResourceImpl(
         userId: String,
         projectId: String,
         templateId: String,
-        version: Long,
+        version: Int,
         useTemplateSettings: Boolean,
         instances: List<TemplateInstanceUpdate>
     ): Result<Boolean> {
