+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/dao/template/TemplateInstanceBaseDao.kt
@@ -218,7 +218,7 @@ class TemplateInstanceBaseDao {
             baseId = id,
             projectId = projectId,
             templateId = templateId,
-            templateVersion = templateVersion.toLong(),
+            templateVersion = templateVersion.toInt(),
             useTemplateSetting = useTemplateSettingsFlag,
             totalItemNum = totalItemNum,
             successItemNum = successItemNum,
