+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/template/v2/PipelineTemplateResourceService.kt
@@ -23,7 +23,7 @@ class PipelineTemplateResourceService @Autowired constructor(
 
     fun getTemplateResourceVersion(
         templateId: String,
-        version: Long
+        version: Int
     ): PipelineTemplateResource? {
         return pipelineTemplateResourceDao.get(
             dslContext = dslContext,
@@ -72,7 +72,7 @@ class PipelineTemplateResourceService @Autowired constructor(
         projectId: String,
         templateId: String,
         status: VersionStatus,
-        version: Long? = null,
+        version: Int? = null,
         versionName: String? = null
     ): PipelineTemplateResource? {
         return pipelineTemplateResourceDao.getLatestRecord(
@@ -124,7 +124,7 @@ class PipelineTemplateResourceService @Autowired constructor(
     fun get(
         projectId: String,
         templateId: String,
-        version: Long
+        version: Int
     ): PipelineTemplateResource {
         return pipelineTemplateResourceDao.get(
             commonCondition = PipelineTemplateResourceCommonCondition(
